chore(scripts): move getWorkspacePaths to utils
