+++ scripts/utils/getWorkspacePaths.mjs
